Handle cert injection properly for authn-k8s (#1789)
* Inject cert using bash script

There is [an issue](https://github.com/kubernetes/kubernetes/issues/89899)
in k8s where messages that are sent via WebSockets to Kubectl Exec are
left hanging without returning a response when using STDIN.

However, if we inject the cert by running a bash script that exits explicitly
then we don't have the hanging issue, and the K8s API will close the socket
upon cert injection success/failure.

* Make exec command timeout configurable

We now enable configuration of the exec command timeout. This
can be done by setting the `KUBECTL_EXEC_COMMAND_TIMEOUT` env var.

* Return 202 instead of 200 for inject_client_cert requests

The "inject_client_cert" request now returns 202 Accepted instead of 200 OK to
indicate that the cert injection has started but not necessarily completed.

* Print cert injection logs upon error

The "inject_client_cert" request now writes the injection logs
to the client container. We can print them when we have an error to
help troubleshooting

* Verify "inject_client_cert" responds with 202

We should verify that the response code of the request
is 202 Accepted in successful cases

* Extract file copy logic to new command class

This logic was inside kubectl_exec.rb which had a lot going
on already. This commit will enhance this logic's readability

* Refactor kubectl_client to kube_client

The latter better tells its purpose as it a k8s client
and not a Kubectl client

* Remove redundant -r flag

* Add description and fix indentation

* Refactor SetFileContentInContainer to CopyTextToFileInContainer

* Add UTs for CopyTextToFileInContainer class

* Implement some PR requests