Align KubeExec with CommandClass style format
- Rename class to the action it performs
- Use dependencies so we can easily mock them
- call function tells the class's story

This commit also moves the error validation of websocket errors
into the `KubeExec` (renamed to `ExecuteCommandInContainer`) class.

Prior to this commit, the KubeExec class didn't verify that
there were no errors in the stream and returned the messages to
the caller. Then, the caller (in this case 'InjectClientCert')
parsed the messages and verified there were no errors. Instead,
it is better to handle the errors in the KubeExec class and raise
an error if one occurred.