Erase cert from every authenticator container
Before each cucumber test of authn-k8s, we delete the cert file
from `/etc/conjur/ssl` so that each test starts with a fresh folder.

However, we currently do this only for pods with the name 'inventory-deployment'.
This is not enough as we have different types of pods with the authenticator container
in it. They all have names that start with 'inventory-' so we can use that as the regex.